refactor(context): split compression strategies
