--- agent_s3/tools/context_management/compression.py
@@ -1,1291 +1,0 @@
-"""
-Context Compression for Agent-S3.
-
-This module provides strategies for compressing context to optimize token usage
-while preserving essential information for LLM processing.
-"""
-
-import re
-import logging
-import hashlib
-from abc import ABC, abstractmethod
-from typing import Dict, List, Any, Optional
-from datetime import datetime
-
-logger = logging.getLogger(__name__)
-
-
-class CompressionStrategy(ABC):
-    """
-    Abstract base class for context compression strategies.
-    """
-
-    @abstractmethod
-    def compress(self, context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Compress the given context.
-
-        Args:
-            context: The context dictionary to compress
-
-        Returns:
-            The compressed context dictionary
-        """
-        pass
-
-    @abstractmethod
-    def decompress(self, compressed_context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Decompress a previously compressed context.
-
-        Args:
-            compressed_context: The compressed context dictionary
-
-        Returns:
-            The decompressed context dictionary
-        """
-        pass
-
-
-class SemanticSummarizer(CompressionStrategy):
-    """
-    Compresses context by providing semantic summaries of file contents.
-    """
-
-    def __init__(
-        self,
-        summarization_threshold: int = 200,
-        preserve_imports: bool = True,
-        preserve_classes: bool = True
-    ):
-        """
-        Initialize the semantic summarizer.
-
-        Args:
-            summarization_threshold: Line count threshold for summarization
-            preserve_imports: Whether to preserve import statements in summaries
-            preserve_classes: Whether to preserve class definitions in summaries
-        """
-        self.summarization_threshold = summarization_threshold
-        self.preserve_imports = preserve_imports
-        self.preserve_classes = preserve_classes
-
-    def compress(self, context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Compress context by summarizing file contents.
-
-        Args:
-            context: The context dictionary to compress
-
-        Returns:
-            The compressed context with summarized file contents
-        """
-        compressed = context.copy()
-
-        # Focus on code_context for summarization
-        if "code_context" in context:
-            compressed_code_context = {}
-
-            for file_path, content in context["code_context"].items():
-                # Count lines to determine if summarization is needed
-                lines = content.split('\n')
-                if len(lines) > self.summarization_threshold:
-                    # Apply summarization
-                    summarized = self._summarize_content(content, file_path)
-                    compressed_code_context[file_path] = summarized
-
-                    # Add metadata about compression
-                    if "compression_metadata" not in compressed:
-                        compressed["compression_metadata"] = {}
-
-                    if "summarized_files" not in compressed["compression_metadata"]:
-                        compressed["compression_metadata"]["summarized_files"] = {}
-
-                    compressed["compression_metadata"]["summarized_files"][file_path] = {
-                        "original_lines": len(lines),
-                        "summarized_lines": len(summarized.split('\n')),
-                        "compression_ratio": len(summarized) / len(content)
-                    }
-                else:
-                    # Keep original content for small files
-                    compressed_code_context[file_path] = content
-
-            compressed["code_context"] = compressed_code_context
-
-            # Add overall compression information
-            if "compression_metadata" in compressed:
-                original_size = sum(len(content) for content in context["code_context"].values())
-                compressed_size = sum(len(content) for content in compressed_code_context.values())
-
-                compressed["compression_metadata"]["overall"] = {
-                    "strategy": "semantic_summarizer",
-                    "original_size": original_size,
-                    "compressed_size": compressed_size,
-                    "compression_ratio": compressed_size / original_size if original_size > 0 else 1.0
-                }
-
-        return compressed
-
-    def decompress(self, compressed_context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        'Decompress' a semantically summarized context.
-
-        Note: Since semantic summarization is lossy, full restoration is not possible.
-        This method adds comprehensive metadata indicating that the context contains
-        summarized content and details about the summarization applied.
-
-        Args:
-            compressed_context: The compressed context dictionary
-
-        Returns:
-            The context with decompression metadata added
-        """
-        decompressed = compressed_context.copy()
-
-        # Ensure decompression metadata section exists
-        if "decompression_metadata" not in decompressed:
-            decompressed["decompression_metadata"] = {}
-
-        # Get information about summarized files
-        summarized_files = {}
-        if "compression_metadata" in compressed_context:
-            if "summarized_files" in compressed_context["compression_metadata"]:
-                summarized_files = compressed_context["compression_metadata"]["summarized_files"]
-
-        # Add detailed decompression metadata
-        decompressed["decompression_metadata"]["semantic_summarization"] = {
-            "timestamp": datetime.now().isoformat(),
-            "status": "completed",
-            "decompression_type": "semantic_summarizer",
-            "note": "Semantic summarization is lossy; original content cannot be fully restored",
-            "summarized_files": list(summarized_files.keys()) if summarized_files else [],
-            "summarization_statistics": {
-                "files_summarized": len(summarized_files),
-                "average_compression_ratio": (
-                    sum(f.get("compression_ratio", 0) for f in summarized_files.values()) /
-                    len(summarized_files) if summarized_files else 0
-                )
-            }
-        }
-
-        return decompressed
-
-    def _summarize_content(self, content: str, file_path: str) -> str:
-        """
-        Summarize the content of a file.
-
-        Args:
-            content: The file content to summarize
-            file_path: Path to the file (used to determine file type)
-
-        Returns:
-            Summarized content
-        """
-        lines = content.split('\n')
-        file_ext = file_path.split('.')[-1].lower() if '.' in file_path else ""
-
-        # Determine language based on file extension
-        language = "unknown"
-        if file_ext in ["py"]:
-            language = "python"
-        elif file_ext in ["js", "jsx", "ts", "tsx"]:
-            language = "javascript"
-        elif file_ext in ["java"]:
-            language = "java"
-        elif file_ext in ["cs"]:
-            language = "csharp"
-
-        # Apply language-specific summarization
-        if language == "python":
-            return self._summarize_python(lines)
-        elif language in ["javascript", "typescript"]:
-            return self._summarize_javascript(lines)
-        elif language == "java":
-            return self._summarize_java(lines)
-        elif language == "csharp":
-            return self._summarize_csharp(lines)
-        else:
-            # Generic summarization for unknown languages
-            return self._summarize_generic(lines)
-
-    def _summarize_python(self, lines: List[str]) -> str:
-        """Summarize Python code."""
-        summary_lines = []
-
-        # Track state
-        current_indent = 0
-        skipping = False
-
-        # Extract imports
-        imports = []
-        if self.preserve_imports:
-            for line in lines:
-                if re.match(r'^import\s+|^from\s+', line):
-                    imports.append(line)
-
-        # Extract class and function definitions
-        for i, line in enumerate(lines):
-            # Check indentation
-            indent = len(line) - len(line.lstrip())
-
-            # Reset skipping if we went back to a lower indent level
-            if indent <= current_indent and skipping:
-                skipping = False
-
-            # Handle class definitions
-            if re.match(r'^\s*class\s+', line):
-                current_indent = indent
-                if self.preserve_classes:
-                    summary_lines.append(line)
-                else:
-                    # Just add the class signature with a comment
-                    class_name = re.search(r'class\s+(\w+)', line).group(1)
-                    summary_lines.append(f"{' ' * indent}class {class_name}: # Summarized")
-                continue
-
-            # Handle function definitions
-            if re.match(r'^\s*def\s+', line):
-                current_indent = indent
-
-                # Add function signature
-                fn_match = re.search(r'def\s+(\w+)\s*\((.*?)\)', line)
-                if fn_match:
-                    fn_name, params = fn_match.groups()
-
-                    # Check if this is a method in a class
-                    is_method = indent > 0
-
-                    # Include the function definition but mark it as summarized
-                    if is_method:
-                        # For methods, include the decorator if it's there
-                        if i > 0 and re.match(r'^\s*@', lines[i-1]):
-                            summary_lines.append(lines[i-1])
-                        summary_lines.append(line)
-
-                        # Skip the function body
-                        skipping = True
-                    else:
-                        # For standalone functions, just add a simplified signature
-                        summary_lines.append(f"{' ' * indent}def {fn_name}(...): # Summarized")
-
-                        # Skip the function body
-                        skipping = True
-                continue
-
-            # Include important comments and docstrings
-            if re.match(r'^\s*"""', line) or re.match(r"^\s*'''", line) or re.match(r'^\s*#', line):
-                if not skipping:
-                    summary_lines.append(line)
-                continue
-
-            # Include non-function, non-class code if not skipping
-            if not skipping:
-                summary_lines.append(line)
-
-        # Combine imports and code summary
-        if imports:
-            return '\n'.join(imports + ['', '# Summarized Content:'] + summary_lines)
-        else:
-            return '\n'.join(['# Summarized Content:'] + summary_lines)
-
-    def _summarize_javascript(self, lines: List[str]) -> str:
-        """Summarize JavaScript/TypeScript code."""
-        summary_lines = []
-
-        # Track state
-        brace_depth = 0
-        skipping = False
-
-        # Extract imports
-        imports = []
-        if self.preserve_imports:
-            for line in lines:
-                if re.match(r'^import\s+|^const\s+.*\s*=\s*require\(', line):
-                    imports.append(line)
-
-        # Extract class and function definitions
-        for i, line in enumerate(lines):
-            # Update brace depth
-            brace_depth += line.count('{') - line.count('}')
-
-            # Reset skipping if we went back to a lower brace depth
-            if brace_depth == 0:
-                skipping = False
-
-            # Handle class definitions
-            if re.search(r'class\s+\w+', line):
-                if self.preserve_classes:
-                    summary_lines.append(line)
-                else:
-                    # Just add the class signature with a comment
-                    class_name = re.search(r'class\s+(\w+)', line).group(1)
-                    summary_lines.append(f"class {class_name} {{ // Summarized")
-
-                # Skip the class body
-                if not self.preserve_classes and '{' in line:
-                    skipping = True
-                continue
-
-            # Handle function definitions
-            if (
-                re.search(r'function\s+\w+\s*\(|^\s*\w+\s*\([^)]*\)\s*{|^\s*\w+\s*:\s*function', line)
-                or re.search(r'const\s+\w+\s*=\s*\([^)]*\)\s*=>|^\s*\w+\s*=\s*\([^)]*\)\s*=>', line)
-            ):
-                # Add function signature
-                summary_lines.append(line.rstrip())
-
-                # Add a comment if this is a named function
-                fn_name = None
-                if re.search(r'function\s+(\w+)', line):
-                    fn_name = re.search(r'function\s+(\w+)', line).group(1)
-                elif re.search(r'(\w+)\s*\(', line):
-                    fn_name = re.search(r'(\w+)\s*\(', line).group(1)
-                elif re.search(r'const\s+(\w+)\s*=', line):
-                    fn_name = re.search(r'const\s+(\w+)\s*=', line).group(1)
-
-                if fn_name and '{' in line:
-                    summary_lines[-1] = summary_lines[-1].replace('{', '{ // Summarized')
-                    skipping = True
-                elif '{' in line:
-                    summary_lines[-1] = summary_lines[-1].replace('{', '{ // Summarized')
-                    skipping = True
-                continue
-
-            # Include important comments and JSDoc
-            if re.match(r'^\s*\/\*\*', line) or re.match(r'^\s*\/\/', line):
-                if not skipping:
-                    summary_lines.append(line)
-                continue
-
-            # Include non-function, non-class code if not skipping
-            if not skipping:
-                summary_lines.append(line)
-
-        # Combine imports and code summary
-        if imports:
-            return '\n'.join(imports + ['', '// Summarized Content:'] + summary_lines)
-        else:
-            return '\n'.join(['// Summarized Content:'] + summary_lines)
-
-    def _summarize_java(self, lines: List[str]) -> str:
-        """Summarize Java code."""
-        # Similar implementation to _summarize_javascript but adapted for Java
-        summary_lines = []
-
-        # Track state
-        brace_depth = 0
-        skipping = False
-
-        # Extract imports
-        imports = []
-        if self.preserve_imports:
-            for line in lines:
-                if re.match(r'^import\s+', line):
-                    imports.append(line)
-                elif re.match(r'^package\s+', line):
-                    imports.append(line)
-
-        # Extract class and method definitions
-        for line in lines:
-            # Update brace depth
-            brace_depth += line.count('{') - line.count('}')
-
-            # Reset skipping if we went back to a lower brace depth
-            if brace_depth == 0:
-                skipping = False
-
-            # Handle class definitions
-            if re.search(r'(public|private|protected)?\s*class\s+\w+', line):
-                if self.preserve_classes:
-                    summary_lines.append(line)
-                else:
-                    # Just add the class signature with a comment
-                    class_match = re.search(r'(?:public|private|protected)?\s*class\s+(\w+)', line)
-                    if class_match:
-                        class_name = class_match.group(1)
-                        summary_lines.append(f"public class {class_name} {{ // Summarized")
-
-                # Skip the class body
-                if not self.preserve_classes and '{' in line:
-                    skipping = True
-                continue
-
-            # Handle method definitions
-            if re.search(r'(public|private|protected)?\s+\w+\s+\w+\s*\([^)]*\)', line):
-
-                # Add method signature
-                summary_lines.append(line.rstrip())
-
-                # Add a comment if this is a method
-                if '{' in line:
-                    summary_lines[-1] = summary_lines[-1].replace('{', '{ // Summarized')
-                    skipping = True
-                continue
-
-            # Include important comments and JavaDoc
-            if re.match(r'^\s*\/\*\*', line) or re.match(r'^\s*\/\/', line):
-                if not skipping:
-                    summary_lines.append(line)
-                continue
-
-            # Include non-method, non-class code if not skipping
-            if not skipping:
-                summary_lines.append(line)
-
-        # Combine imports and code summary
-        if imports:
-            return '\n'.join(imports + ['', '// Summarized Content:'] + summary_lines)
-        else:
-            return '\n'.join(['// Summarized Content:'] + summary_lines)
-
-    def _summarize_csharp(self, lines: List[str]) -> str:
-        """Summarize C# code."""
-        # Similar to _summarize_java but adapted for C#
-        summary_lines = []
-
-        # Track state
-        brace_depth = 0
-        skipping = False
-
-        # Extract imports
-        imports = []
-        if self.preserve_imports:
-            for line in lines:
-                if re.match(r'^using\s+', line):
-                    imports.append(line)
-                elif re.match(r'^namespace\s+', line):
-                    imports.append(line)
-
-        # Extract class and method definitions
-        for line in lines:
-            # Update brace depth
-            brace_depth += line.count('{') - line.count('}')
-
-            # Reset skipping if we went back to a lower brace depth
-            if brace_depth == 0:
-                skipping = False
-
-            # Handle class definitions
-            if re.search(r'(public|private|protected|internal)?\s*class\s+\w+', line):
-                if self.preserve_classes:
-                    summary_lines.append(line)
-                else:
-                    # Just add the class signature with a comment
-                    class_match = re.search(
-                        r"(?:public|private|protected|internal)?\s*class\s+(\w+)",
-                        line,
-                    )
-                    if class_match:
-                        class_name = class_match.group(1)
-                        summary_lines.append(f"public class {class_name} {{ // Summarized")
-
-                # Skip the class body
-                if not self.preserve_classes and '{' in line:
-                    skipping = True
-                continue
-
-            # Handle method definitions
-            if re.search(r'(public|private|protected|internal)?\s+\w+\s+\w+\s*\([^)]*\)', line):
-
-                # Add method signature
-                summary_lines.append(line.rstrip())
-
-                # Add a comment if this is a method
-                if '{' in line:
-                    summary_lines[-1] = summary_lines[-1].replace('{', '{ // Summarized')
-                    skipping = True
-                continue
-
-            # Include important comments and C# Doc
-            if re.match(r'^\s*\/\/\/\s*<', line) or re.match(r'^\s*\/\/', line):
-                if not skipping:
-                    summary_lines.append(line)
-                continue
-
-            # Include non-method, non-class code if not skipping
-            if not skipping:
-                summary_lines.append(line)
-
-        # Combine imports and code summary
-        if imports:
-            return '\n'.join(imports + ['', '// Summarized Content:'] + summary_lines)
-        else:
-            return '\n'.join(['// Summarized Content:'] + summary_lines)
-
-    def _summarize_generic(self, lines: List[str]) -> str:
-        """Generic summarization for unknown languages."""
-        # For unknown languages, include first 10% and last 10% of lines
-        if len(lines) <= self.summarization_threshold:
-            return '\n'.join(lines)
-
-        header_size = max(10, int(len(lines) * 0.1))
-        footer_size = max(10, int(len(lines) * 0.1))
-
-        header = lines[:header_size]
-        footer = lines[-footer_size:]
-
-        # Add a separator between header and footer
-        separator = [
-            '',
-            '// ...',
-            f'// [Content summarized: {len(lines) - header_size - footer_size} lines omitted]',
-            '// ...',
-            ''
-        ]
-
-        return '\n'.join(header + separator + footer)
-
-
-class KeyInfoExtractor(CompressionStrategy):
-    """
-    Extracts and preserves only key information from context.
-    """
-
-    def __init__(
-        self,
-        extraction_patterns: Optional[Dict[str, List[str]]] = None,
-        preserve_structure: bool = True
-    ):
-        """
-        Initialize the key information extractor.
-
-        Args:
-            extraction_patterns: Patterns for extracting key information by language
-            preserve_structure: Whether to preserve the overall structure
-        """
-        self.preserve_structure = preserve_structure
-        self.extraction_patterns = extraction_patterns or {
-            "python": [
-                r'^\s*import\s+.*$',
-                r'^\s*from\s+.*\s+import\s+.*$',
-                r'^\s*class\s+\w+.*:$',
-                r'^\s*def\s+\w+\s*\(.*\):$',
-                r'^\s*@.*$',
-                r'^\s*""".*?"""$'
-            ],
-            "javascript": [
-                r'^\s*import\s+.*$',
-                r'^\s*export\s+.*$',
-                r'^\s*class\s+\w+.*{$',
-                r'^\s*function\s+\w+\s*\(.*\)\s*{$',
-                r'^\s*const\s+\w+\s*=\s*\(.*\)\s*=>.*$',
-                r'^\s*\/\*\*.*?\*\/$'
-            ],
-            "generic": [
-                r'^\s*function\s+\w+',
-                r'^\s*class\s+\w+',
-                r'^\s*\w+\s*\(',
-                r'^\s*\/\/\s*\w+'
-            ]
-        }
-
-    def compress(self, context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Compress context by extracting key information.
-
-        Args:
-            context: The context dictionary to compress
-
-        Returns:
-            The compressed context with only key information
-        """
-        compressed = context.copy()
-
-        # Focus on code_context for key info extraction
-        if "code_context" in context:
-            compressed_code_context = {}
-
-            for file_path, content in context["code_context"].items():
-                # Extract key information from content
-                extracted = self._extract_key_info(content, file_path)
-                compressed_code_context[file_path] = extracted
-
-            compressed["code_context"] = compressed_code_context
-
-            # Add compression metadata
-            if "compression_metadata" not in compressed:
-                compressed["compression_metadata"] = {}
-
-            original_size = sum(len(content) for content in context["code_context"].values())
-            compressed_size = sum(len(content) for content in compressed_code_context.values())
-
-            compressed["compression_metadata"]["overall"] = {
-                "strategy": "key_info_extractor",
-                "original_size": original_size,
-                "compressed_size": compressed_size,
-                "compression_ratio": compressed_size / original_size if original_size > 0 else 1.0
-            }
-
-        return compressed
-
-    def decompress(self, compressed_context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        'Decompress' a context with only key information.
-
-        Note: Since key information extraction is lossy, full restoration is not possible.
-        This method adds detailed metadata indicating that the context contains
-        only key information and the nature of the extraction performed.
-
-        Args:
-            compressed_context: The compressed context dictionary
-
-        Returns:
-            The context with decompression metadata added
-        """
-        decompressed = compressed_context.copy()
-
-        # Ensure decompression metadata section exists
-        if "decompression_metadata" not in decompressed:
-            decompressed["decompression_metadata"] = {}
-
-        # Extract information about the compression
-        extracted_files = []
-        if "code_context" in compressed_context:
-            extracted_files = list(compressed_context["code_context"].keys())
-
-        compression_stats = {}
-        if "compression_metadata" in compressed_context and "overall" in compressed_context["compression_metadata"]:
-            compression_stats = compressed_context["compression_metadata"]["overall"]
-
-        # Add comprehensive decompression metadata
-        decompressed["decompression_metadata"]["key_info_extraction"] = {
-            "timestamp": datetime.now().isoformat(),
-            "status": "completed",
-            "decompression_type": "key_info_extractor",
-            "note": "Key information extraction is lossy; original content cannot be fully restored",
-            "extracted_files": extracted_files,
-            "extraction_info": {
-                "files_processed": len(extracted_files),
-                "original_size": compression_stats.get("original_size", 0),
-                "compressed_size": compression_stats.get("compressed_size", 0),
-                "compression_ratio": compression_stats.get("compression_ratio", 1.0),
-                "patterns_used": [p for patterns in self.extraction_patterns.values() for p in patterns]
-            }
-        }
-
-        return decompressed
-
-    def _extract_key_info(self, content: str, file_path: str) -> str:
-        """
-        Extract key information from file content.
-
-        Args:
-            content: The file content to extract from
-            file_path: Path to the file (used to determine file type)
-
-        Returns:
-            Content with only key information
-        """
-        lines = content.split('\n')
-        file_ext = file_path.split('.')[-1].lower() if '.' in file_path else ""
-
-        # Determine language based on file extension
-        language = "generic"
-        if file_ext in ["py"]:
-            language = "python"
-        elif file_ext in ["js", "jsx", "ts", "tsx"]:
-            language = "javascript"
-
-        # Get patterns for this language
-        patterns = self.extraction_patterns.get(language, self.extraction_patterns["generic"])
-
-        # Extract matching lines
-        extracted_lines = []
-        structure_indicators = []
-
-        for i, line in enumerate(lines):
-            # Check if the line matches any pattern
-            if any(re.search(pattern, line, re.DOTALL) for pattern in patterns):
-                extracted_lines.append((i, line))
-            elif line.strip() and self.preserve_structure:
-                # Record indentation level for structure preservation
-                indent = len(line) - len(line.lstrip())
-                if indent > 0:
-                    structure_indicators.append((i, indent))
-
-        # If preserving structure, add indicator lines
-        if self.preserve_structure:
-            # Group structure indicators by indentation level
-            indent_groups = {}
-            for i, indent in structure_indicators:
-                if indent not in indent_groups:
-                    indent_groups[indent] = []
-                indent_groups[indent].append(i)
-
-            # Add indicator lines with original indentation
-            for indent, positions in indent_groups.items():
-                for group in self._group_consecutive(positions):
-                    if group and len(group) > 2:  # Only add for larger blocks
-                        mid_pos = group[len(group) // 2]
-                        extracted_lines.append((mid_pos, ' ' * indent + "// ..."))
-
-        # Sort by original line number
-        extracted_lines.sort()
-
-        # Build the final extracted content
-        extracted_content = []
-        last_line = -1
-
-        for i, line in extracted_lines:
-            # Add separator if there's a gap
-            if last_line >= 0 and i - last_line > 1:
-                extracted_content.append("// ...")
-
-            extracted_content.append(line)
-            last_line = i
-
-        # Add header with extraction information
-        header = [
-            "// Key Information Extract",
-            f"// Original file: {file_path}",
-            f"// Extraction patterns: {language}",
-            f"// Original size: {len(lines)} lines",
-            f"// Extracted size: {len(extracted_content)} elements",
-            "// Note: This is a compressed representation with only key elements",
-            ""
-        ]
-
-        return '\n'.join(header + extracted_content)
-
-    def _group_consecutive(self, numbers: List[int]) -> List[List[int]]:
-        """Group consecutive numbers into sublists."""
-        groups = []
-        if not numbers:
-            return groups
-
-        current_group = [numbers[0]]
-
-        for i in range(1, len(numbers)):
-            if numbers[i] == numbers[i-1] + 1:
-                current_group.append(numbers[i])
-            else:
-                groups.append(current_group)
-                current_group = [numbers[i]]
-
-        if current_group:
-            groups.append(current_group)
-
-        return groups
-
-
-class ReferenceCompressor(CompressionStrategy):
-    """
-    Compresses context by replacing repeated code with references.
-    """
-
-    def __init__(
-        self,
-        min_pattern_length: int = 10,
-        similarity_threshold: float = 0.8
-    ):
-        """
-        Initialize the reference compressor.
-
-        Args:
-            min_pattern_length: Minimum length of patterns to identify
-            similarity_threshold: Threshold for identifying similar code blocks
-        """
-        self.min_pattern_length = min_pattern_length
-        self.similarity_threshold = similarity_threshold
-        self.reference_map = {}
-
-    def compress(self, context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Compress context by replacing repeated code with references.
-
-        Args:
-            context: The context dictionary to compress
-
-        Returns:
-            The compressed context with referenced content
-        """
-        compressed = context.copy()
-
-        # Reset reference map
-        self.reference_map = {}
-
-        # Focus on code_context for reference compression
-        if "code_context" in context:
-            # First pass: identify repeating patterns
-            self._identify_patterns(context["code_context"])
-
-            # Second pass: replace patterns with references
-            compressed_code_context = {}
-
-            for file_path, content in context["code_context"].items():
-                if not content:
-                    compressed_code_context[file_path] = content
-                    continue
-
-                # Apply reference compression
-                compressed_content = self._apply_references(content)
-                compressed_code_context[file_path] = compressed_content
-
-            compressed["code_context"] = compressed_code_context
-
-            # Add reference map to context
-            if "compression_metadata" not in compressed:
-                compressed["compression_metadata"] = {}
-
-            compressed["compression_metadata"]["reference_map"] = self.reference_map
-
-            # Add overall compression information
-            original_size = sum(len(content) for content in context["code_context"].values())
-            compressed_size = sum(len(content) for content in compressed_code_context.values())
-
-            compressed["compression_metadata"]["overall"] = {
-                "strategy": "reference_compressor",
-                "original_size": original_size,
-                "compressed_size": compressed_size,
-                "compression_ratio": compressed_size / original_size if original_size > 0 else 1.0
-            }
-
-        return compressed
-
-    def decompress(self, compressed_context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Decompress a reference-compressed context.
-
-        Args:
-            compressed_context: The compressed context dictionary
-
-        Returns:
-            The decompressed context with references expanded
-        """
-        decompressed = compressed_context.copy()
-
-        # Extract reference map
-        if "compression_metadata" in compressed_context and "reference_map" in compressed_context["compression_metadata"]:
-            reference_map = compressed_context["compression_metadata"]["reference_map"]
-
-            # Decompress code_context
-            if "code_context" in compressed_context:
-                decompressed_code_context = {}
-
-                for file_path, content in compressed_context["code_context"].items():
-                    # Expand references
-                    decompressed_content = self._expand_references(content, reference_map)
-                    decompressed_code_context[file_path] = decompressed_content
-
-                decompressed["code_context"] = decompressed_code_context
-
-                # Add decompression metadata
-                if "decompression_metadata" not in decompressed:
-                    decompressed["decompression_metadata"] = {}
-
-                decompressed["decompression_metadata"]["reference_decompression"] = {
-                    "references_expanded": len(reference_map)
-                }
-
-        return decompressed
-
-    def _identify_patterns(self, code_context: Dict[str, str]) -> None:
-        """
-        Identify repeating patterns in code context.
-
-        Args:
-            code_context: Dictionary mapping file paths to content
-        """
-        # Extract all content
-        all_content = '\n'.join(content for content in code_context.values() if content)
-
-        # Split into lines
-        lines = all_content.split('\n')
-
-        # Initialize pattern tracking
-        potential_patterns = {}
-
-        # Process lines in chunks
-        for i in range(len(lines) - self.min_pattern_length + 1):
-            chunk = '\n'.join(lines[i:i + self.min_pattern_length])
-
-            # Skip if too short
-            if len(chunk) < self.min_pattern_length * 5:  # Average 5 chars per line minimum
-                continue
-
-            # Hash the chunk for quicker comparison
-            chunk_hash = hashlib.sha256(chunk.encode()).hexdigest()
-
-            if chunk_hash in potential_patterns:
-                potential_patterns[chunk_hash]["count"] += 1
-            else:
-                potential_patterns[chunk_hash] = {
-                    "content": chunk,
-                    "count": 1
-                }
-
-        # Adaptively reduce pattern size for smaller content
-        if not potential_patterns or all(v["count"] <= 1 for v in potential_patterns.values()):
-            # Determine pattern size based on content length
-            adaptive_length = max(2, min(5, len(lines) // 10))
-
-            # Use adaptive pattern length
-            for i in range(len(lines) - adaptive_length + 1):
-                chunk = '\n'.join(lines[i:i + adaptive_length])
-                if len(chunk) < 10:  # Skip very small chunks
-                    continue
-
-                chunk_hash = hashlib.sha256(chunk.encode()).hexdigest()
-                if chunk_hash in potential_patterns:
-                    potential_patterns[chunk_hash]["count"] += 1
-                else:
-                    potential_patterns[chunk_hash] = {
-                        "content": chunk,
-                        "count": 1
-                    }
-
-        # Filter to patterns that appear multiple times
-        repeating_patterns = {
-            k: v for k, v in potential_patterns.items()
-            if v["count"] > 1 and len(v["content"]) > 10
-        }
-
-        # Convert to reference map
-        ref_id = 1
-        for pattern_hash, pattern_info in repeating_patterns.items():
-            ref_key = f"@REF{ref_id}@"
-            self.reference_map[ref_key] = pattern_info["content"]
-            ref_id += 1
-
-    def _apply_references(self, content: str) -> str:
-        """
-        Replace repeated patterns with references.
-
-        Args:
-            content: Original content
-
-        Returns:
-            Content with patterns replaced by references
-        """
-        if not self.reference_map or not content:
-            return content
-
-        compressed = content
-
-        for ref_key, pattern in self.reference_map.items():
-            if pattern in compressed:
-                # Replace the pattern with a reference
-                replacement = f"\n// {ref_key} - Reference to a repeated pattern\n"
-                compressed = compressed.replace(pattern, replacement)
-
-        # Add header with reference information
-        if compressed != content:
-            header = [
-                "// Reference-Compressed Content",
-                "// This file contains references to repeated patterns",
-                "// References are marked with @REFn@ tags",
-                ""
-            ]
-            compressed = '\n'.join(header) + compressed
-
-        return compressed
-
-    def _expand_references(self, content: str, reference_map: Dict[str, str]) -> str:
-        """
-        Expand references back to original content.
-
-        Args:
-            content: Content with references
-            reference_map: Map of reference keys to original content
-
-        Returns:
-            Content with references expanded
-        """
-        if not reference_map or not content:
-            return content
-
-        decompressed = content
-
-        for ref_key, pattern in reference_map.items():
-            ref_marker = f"// {ref_key} - Reference to a repeated pattern"
-            if ref_marker in decompressed:
-                # Replace the reference with the original pattern
-                decompressed = decompressed.replace(ref_marker, pattern)
-
-        # Remove reference header if present
-        header_lines = [
-            "// Reference-Compressed Content",
-            "// This file contains references to repeated patterns",
-            "// References are marked with @REFn@ tags",
-            ""
-        ]
-        for line in header_lines:
-            if decompressed.startswith(line):
-                decompressed = decompressed[len(line)+1:]  # +1 for newline
-
-        return decompressed
-
-
-class CompressionManager:
-    """
-    Manages context compression using various strategies.
-    """
-
-    def __init__(
-        self,
-        compression_threshold: int = 32000,
-        min_compression_ratio: float = 0.7,
-        strategies: Optional[List[CompressionStrategy]] = None
-    ):
-        """
-        Initialize the compression manager.
-
-        Args:
-            compression_threshold: Token threshold to trigger compression
-            min_compression_ratio: Minimum compression ratio to accept
-            strategies: List of compression strategies to use
-        """
-        self.compression_threshold = compression_threshold
-        self.min_compression_ratio = min_compression_ratio
-        self.strategies = strategies or [
-            SemanticSummarizer(),
-            KeyInfoExtractor(),
-            ReferenceCompressor()
-        ]
-
-    def need_compression(self, context: Dict[str, Any], token_count: Optional[int] = None) -> bool:
-        """
-        Determine if context needs compression.
-
-        Args:
-            context: The context dictionary
-            token_count: Optional pre-calculated token count
-
-        Returns:
-            True if compression is needed, False otherwise
-        """
-        if token_count is not None:
-            return token_count > self.compression_threshold
-
-        # Calculate total characters in all context
-        char_count = 0
-
-        # Calculate from code_context (most common case)
-        if "code_context" in context:
-            char_count += sum(len(content) for content in context["code_context"].values())
-
-        # Include all other context keys as well for more accurate assessment
-        for key, value in context.items():
-            if key != "code_context" and key != "compression_metadata":
-                # Use string representation for size calculation
-                char_count += len(str(value))
-
-        # Rough estimate: 1 token ~= 4 chars
-        return char_count / 4 > self.compression_threshold
-
-    def compress(
-        self,
-        context: Dict[str, Any],
-        strategy_names: Optional[List[str]] = None
-    ) -> Dict[str, Any]:
-        """
-        Compress context using specified or all strategies.
-
-        Args:
-            context: The context dictionary to compress
-            strategy_names: Optional list of strategy names to use
-
-        Returns:
-            The compressed context
-        """
-        # In tests, always compress if a strategy is explicitly specified
-        force_compression = strategy_names is not None and len(strategy_names) > 0
-
-        # If no compression needed and not forced, return original
-        if not force_compression and not self.need_compression(context):
-            return context.copy()
-
-        # Determine which strategies to use
-        active_strategies = []
-        if strategy_names:
-            for name in strategy_names:
-                for strategy in self.strategies:
-                    if strategy.__class__.__name__.lower() == name.lower():
-                        active_strategies.append(strategy)
-        else:
-            active_strategies = self.strategies
-
-        # Try each strategy and pick the best one
-        best_compression = None
-        best_ratio = 1.0
-
-        for strategy in active_strategies:
-            try:
-                compressed = strategy.compress(context)
-
-                # Calculate and add proper compression metadata (replacing dummy values with actual calculations)
-                if "compression_metadata" not in compressed:
-                    # Calculate original and compressed sizes
-                    original_size = 0
-                    compressed_size = 0
-
-                    # Sum size of code_context in both original and compressed
-                    if "code_context" in context and "code_context" in compressed:
-                        original_size +
-                            = sum(len(content) for content in context["code_context"].values())                        compressed_size +
-                                                                                            = sum(len(content) for content in compressed["code_context"].values())
-                    # Sum size of other context elements
-                    for key in context:
-                        if key != "code_context" and key != "compression_metadata":
-                            original_size += len(str(context[key]))
-
-                    for key in compressed:
-                        if key != "code_context" and key != "compression_metadata":
-                            compressed_size += len(str(compressed[key]))
-
-                    # Calculate ratio (avoid division by zero)
-                    ratio = compressed_size / original_size if original_size > 0 else 1.0
-
-                    # Add metadata
-                    compressed["compression_metadata"] = {
-                        "overall": {
-                            "strategy": strategy.__class__.__name__,
-                            "original_size": original_size,
-                            "compressed_size": compressed_size,
-                            "compression_ratio": ratio
-                        }
-                    }
-
-                # Calculate compression ratio
-                if "compression_metadata" in compressed and "overall" in compressed["compression_metadata"]:
-                    ratio = compressed["compression_metadata"]["overall"]["compression_ratio"]
-
-                    if ratio < best_ratio:
-                        best_compression = compressed
-                        best_ratio = ratio
-            except Exception as e:
-                logger.warning("%s", Compression strategy {strategy.__class__.__name__} failed: {e})
-
-        # Return the best compression if it meets the minimum ratio or if compression is forced
-        if best_compression and (force_compression or best_ratio <= self.min_compression_ratio):
-            return best_compression
-
-        # If compression is forced but no strategy worked well enough, implement proper compression
-        # (replacing dummy hardcoded values with actual calculations)
-        if force_compression and not best_compression and active_strategies:
-            # Create a proper compressed version using first strategy
-            strategy = active_strategies[0]
-            try:
-                # Try direct compression
-                compressed = strategy.compress(context)
-
-                # Ensure it has compression metadata
-                if "compression_metadata" not in compressed:
-                    # Calculate sizes
-                    original_size = 0
-                    compressed_size = 0
-
-                    # Sum code_context sizes
-                    if "code_context" in context and "code_context" in compressed:
-                        original_size +
-                            = sum(len(content) for content in context["code_context"].values())                        compressed_size +
-                                                                                            = sum(len(content) for content in compressed["code_context"].values())
-                    # Sum other elements
-                    for key in context:
-                        if key != "code_context" and key != "compression_metadata":
-                            original_size += len(str(context[key]))
-
-                    for key in compressed:
-                        if key != "code_context" and key != "compression_metadata":
-                            compressed_size += len(str(compressed[key]))
-
-                    # Calculate ratio
-                    ratio = compressed_size / original_size if original_size > 0 else 0.9
-
-                    # Add metadata
-                    compressed["compression_metadata"] = {
-                        "overall": {
-                            "strategy": strategy.__class__.__name__,
-                            "original_size": original_size,
-                            "compressed_size": compressed_size,
-                            "compression_ratio": ratio
-                        }
-                    }
-
-                return compressed
-            except Exception as e:
-                logger.warning("%s", Forced compression with {strategy.__class__.__name__} failed: {e})
-                # If direct compression fails, create minimal metadata
-                compressed = context.copy()
-                if "compression_metadata" not in compressed:
-                    compressed["compression_metadata"] = {
-                        "overall": {
-                            "strategy": strategy.__class__.__name__,
-                            "status": "minimal_metadata_only",
-                            "original_size": len(str(context)),
-                            "compressed_size": len(str(compressed)),
-                            "compression_ratio": 0.95  # Minimal compression
-                        }
-                    }
-                return compressed
-
-        # If no strategy worked well enough, return original
-        return context.copy()
-
-    def decompress(self, compressed_context: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Decompress context using the appropriate strategy.
-
-        Args:
-            compressed_context: The compressed context dictionary
-
-        Returns:
-            The decompressed context with appropriate decompression metadata
-        """
-        result = compressed_context.copy()
-
-        # Check if this context was compressed
-        if "compression_metadata" in compressed_context and "overall" in compressed_context["compression_metadata"]:
-            strategy_name = compressed_context["compression_metadata"]["overall"]["strategy"]
-
-            # Find the appropriate strategy
-            strategy_found = False
-            for strategy in self.strategies:
-                if strategy.__class__.__name__.lower() == strategy_name.lower():
-                    strategy_found = True
-                    result = strategy.decompress(compressed_context)
-
-                    # Ensure all decompressed results contain metadata about decompression
-                    if "decompression_metadata" not in result:
-                        result["decompression_metadata"] = {}
-
-                    # Add standardized decompression metadata
-                    result["decompression_metadata"][f"{strategy.__class__.__name__.lower()}_decompression"] = {
-                        "timestamp": datetime.now().isoformat(),
-                        "strategy_used": strategy.__class__.__name__,
-                        "compression_ratio": compressed_context["compression_metadata"]["overall"].get("compression_ratio", 1.0)
-                    }
-
-                    break  # Found and used the appropriate strategy
-
-            # Strategy specified in metadata but not available in registered strategies
-            if not strategy_found:
-                if "decompression_metadata" not in result:
-                    result["decompression_metadata"] = {}
-
-                result["decompression_metadata"]["decompression_error"] = {
-                    "timestamp": datetime.now().isoformat(),
-                    "error": f"Strategy '{strategy_name}' not found in available strategies",
-                    "available_strategies": [s.__class__.__name__ for s in self.strategies]
-                }
-        else:
-            # Not a compressed context, add metadata indicating this
-            if "decompression_metadata" not in result:
-                result["decompression_metadata"] = {}
-
-            result["decompression_metadata"]["decompression_skipped"] = {
-                "timestamp": datetime.now().isoformat(),
-                "reason": "Content was not compressed or missing compression metadata"
-            }
-
-        return result
-
-    def get_available_strategies(self) -> List[str]:
-        """
-        Get a list of available compression strategy names.
-
-        Returns:
-            List of strategy names
-        """
-        return [strategy.__class__.__name__ for strategy in self.strategies] + \
-               [strategy.__class__.__name__.lower() for strategy in self.strategies]
